requirements.txt
# Core dependencies
boto3>=1.34.0
botocore>=1.34.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Image capture
opencv-python>=4.9.0.80

# HTTP utilities (already in stdlib, kept for explicit listing)
# urllib.request — stdlib

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-mock>=3.12.0

# Code quality
black>=24.0.0
ruff>=0.3.0

# Optional: web dashboard backend (if running with Flask API)
flask>=3.0.0

# Optional: AWS deployment
# aws-sam-cli — install separately
